Rename ddp_planes → parallel_state, ddp_basic_building_blocks → parallel_primitives
Generalizes naming away from "DDP" since the primitives don't assume the
specific torch.distributed-DDP topology — they hold for FSDP / ZeRO / any
SPMD setup with a rendezvous-on-collective contract.

- weightslab/components/parallel_primitives.py (was ddp_basic_building_blocks)
- weightslab/components/parallel_state.py       (was ddp_planes)
- docs/ddp_design.md                            (was components/ddp_design_notes.md)

Updated import sites (4): global_monitoring.py, dataframe_manager.py,
parallel_primitives.py (self-ref), tests/test_ddp_primitives.py.

Also: scenario_lr_batch_propagate threshold loosened from
`(expected + rank0_only) / 2` to `rank0_only + 1`. The old midpoint sat
right at the noise band; under drop_last=False with mid-iter batch-size
transitions the observed rate floats 13–14 samples/step, occasionally
tripping at exactly 13.75 < 14. New threshold cleanly distinguishes
"both ranks doubled" from "only rank-0 doubled" (rate ~12).
